Renamed rt commande to rct.
This includes:
- renaming the tool iteself
- renaming the python module
- renaming the 'cc' command to 'cat-cert'
